First attempt at implementing changeset dumps.
Changeset-only dumps are pretty useful for doing some analysis
which doesn't require downloading the whole planet, and the old
planet dump script supported it, so it seems like a good idea to
try and support it here too.

This adds a "filter" applied to `xml_writer` which drops all
elements except for changesets, which is pretty much what we want.
Hopefully the overhead of the extra synchronisation with the dump
thread won't be too high.