Add support for changeset discussions in the dump.
This adds support for an *additional* dump of changesets which
includes discussions. This will be the new format and deprecates
the previous format, but we'll allow a few months after the
announcement for the previous format to be sunsetted.

The format is the same as getting changesets with discussions from
the XML API.