Initialise 'zero' entries properly.
Previously, initial entries were initialized to zero - a value for
element IDs which doesn't exist in the database. However, when a
database is loaded using Osmosis, it does create some entries with
ID=0. Therefore, it's better to just initialise to a negative
number and ensure that negative elements are never processed.