Made generator name configurable via the command line.
The generator name appears in the output files (both XML and PBF)
which makes it annoying to test, as a bumped version invalidates
all the static test comparison files. To fix it, this adds an
option to override the `PACKAGE_STRING` so that the tests can use
a stable value. It is not recommended for other use, as having
the current version in the dump file is useful for tracing back
any errors or bug reports.