Saner default for handling existing disk DBs.
Previously, the default behaviour was to try to resume processing
using any existing data that was present. This had the unintended
side effect of silently using old data if it wasn't cleaned up.

I'm sure this being the default made sense at the time, but it
doesn't any more and it confused me greatly. Therefore, the new
default is to remove any existing data and start from scratch. The
old behaviour is available using the `--resume` flag.