Hack to prevent oversized relations.
PBF format requires that the uncompressed size of a block is no
larger than 32MB. Calculating the size of a block has a cost, so
it is only done periodically (for relations, every 2000 elements).

The issue arises because 2000 relations manage to be larger than
24MB (the difference between the cut-off and the max size). The
quick hack fix for this is to reduce the periodic check down to
every 500 relations.

This is obviously not a real fix; for that, we'll need to be able
to calculate the size of each element before adding it to the
block, including the size of any new strings added to the string
table. And, moreover, do that in an efficient manner.

However, this fix should be enough to allow us to get the next
planet out.